Allow crosshair size below 1 (0.25/0.5/0.75)
The size steps are non-uniform now (quarter-steps below 1, whole numbers
1-6 above), which the previous IntSliderButton's linear [min, max]
mapping can't represent. Replaced with DiscreteSliderButton<T>, mapping
slider position to an index into a fixed ordered value list instead of a
number range - IntSliderButton had no other callers, so it's deleted
rather than left unused.

ClientConfig#crosshairPixelSize is now a float, and CrosshairGrid#render
computes cell positions in float, rounding only at the final fill(...)
call. Each cell still fills at least 1 physical pixel so it doesn't just
disappear at very small sizes - at 0.25, adjacent grid cells naturally
round to the same screen pixel and the shape compresses into a blob/dot,
which is the expected trade-off of a crosshair only a couple of pixels
wide, not a bug.

+++ mod/src/main/java/com/tntsallin1client/config/ClientConfig.java
@@ -68,7 +68,9 @@ public class ClientConfig {
 	public CrosshairMode crosshairMode = CrosshairMode.PRESET;
 	public CrosshairPreset crosshairPreset = CrosshairPreset.SMALLER;
 	public boolean[][] crosshairCustomGrid = CrosshairGrid.empty();
-	public int crosshairPixelSize = 2;
+	// Float, not int: the size list has non-uniform steps (0.25/0.5/0.75 below
+	// 1, then whole numbers 1-6 above), see DiscreteSliderButton/CrosshairOptionsScreen.
+	public float crosshairPixelSize = 2.0F;
 	public boolean crosshairIgnoreGuiScale = false;
 	public boolean crosshairTargetColorEnabled = false;
 	public int crosshairTargetColor = 0xFFFF5555;
